Pass build id in ConcourseAPI::get_build
The build object that the concourse API returns has the properties `id`
and `name`. The first is a number while the second is string number.
These are not the same.
Previously, the endpoint GET /builds/{id} was mistakenly accepting the
build name. Now this is enforced by type to be the id instead.