Add revision comment to patch when the CI build has completed
There is a change in the struct that that a `Worker` receives. It now
receives the repo id, the patch id and the profile. The worker now is
able to trigger the CI pipeline and add a revision comment to the patch.
In addition, the Runtime methods were a bit simplified due to this
change.