Load Concourse configuration from file
The Concourse configuration found in the `.concourse` folder is now simplified.
The concourse git resource is no longer used due to not be able to fetch
correctly the patch revision.

The same configuration was present in the create_pipeline method of
ConcourseAPI struct. In order to avoid duplication, the Concourse
configuration is loaded from the file `.concourse/config.yaml` and passed
as template to that function. Then the proper replacements are done and
then the request to the Concourse API is executed.