Trigger pipeline job build based on the new configuration
There's a different, simpler and more generic approach now that the concourse
git resource is removed from the configuration. The list of jbs in the
pipeline is fetched. We currently assume that a single job will be
present there. We use that job to trigger a build. Then that build is
being watched for completion and then a result is posted as a revision
comment to the patch.