Query for pipeline configuration before creation
During the setup phase, we need to query Concourse in order to find out
if a pipeline related to the repository we want to trigger already
exists. If that's the case then we need to find its version and pass it
in the next call to `ConcourseAPI.create_pipeline_config`. As we found
out, Concourse will either create or update the pipeline after this
call. The version is needed to correctly update an existing pipeline.