Refresh Concourse API token when it expires
The implementation refreshes the token if needed in all methods of
ConcourseAPI. A major change is that the CI trait now has both its
methods to accept a mutable self reference. This causes the borrower
checker to complain in the `run_pipeline` method when self is mut
borrowed by `watch_pipeline_job_build`. Until a solution is found for
this, the body of `watch_pipeline_job_build` is inlined.