Create the CIIntegration abstraction
The worker's responsibility is to find out what type of integrations
exist in the storage repository. Right now, we know that we only
integrate with Concourse but this will later change.
The new CIIntegration struct will implement the `CIObserver` trait and
it will be responsible to execute the pipeline and create patch comments
based on the state changes.