Notify observer when ConcourseCI finishes execution
Finally, after the observer is attached to the observable ConcourseCI,
the latter calls its update method in order to pass the CI result as
soon as the pipeline finishes. The function `CI::run_pipeline` still
returns a `CIResult` wrapped in a `Result` but this is not used by the
`CIIntegration`.